+++ src/View/Helper/SearchForm.php
@@ -39,6 +39,31 @@ class SearchForm extends AbstractHelper
         return $this->form;
     }
 
+    public function getAvailableSearchFields()
+    {
+        $index = $this->searchPage->index();
+        $adapter = $index->adapter();
+
+        $settings = $this->searchPage->settings();
+
+        if (!empty($settings['form']['search_fields'])) {
+            $searchFields = $adapter->getAvailableSearchFields($index);
+            return array_filter($searchFields, function ($searchField) use ($settings) {
+                return in_array($searchField['name'], $settings['form']['search_fields']);
+            });
+        }
+
+        return [];
+    }
+
+    public function getAvailableOperators()
+    {
+        $index = $this->searchPage->index();
+        $adapter = $index->adapter();
+
+        return $adapter->getAvailableOperators($index);
+    }
+
     protected function getFormPartial()
     {
         $formAdapter = $this->searchPage->formAdapter();